+++ internal/policy/policy.go
@@ -73,6 +73,7 @@ type Condition struct {
 	CertPubKeyAlgIn          []string `yaml:"cert_pubkey_alg_in,omitempty" json:"cert_pubkey_alg_in,omitempty"`
 	CertCAWithoutKeyCertSign *bool    `yaml:"cert_ca_without_keycertsign,omitempty" json:"cert_ca_without_keycertsign,omitempty"`
 	CertKeyMismatch          *bool    `yaml:"cert_key_mismatch,omitempty" json:"cert_key_mismatch,omitempty"`
+	CertChainBroken          *bool    `yaml:"cert_chain_broken,omitempty" json:"cert_chain_broken,omitempty"`
 
 	DuplicateLabel *bool `yaml:"duplicate_label,omitempty" json:"duplicate_label,omitempty"`
 	Orphan         *bool `yaml:"orphan,omitempty" json:"orphan,omitempty"`
@@ -103,7 +104,7 @@ func (c *Condition) empty() bool {
 		c.CertValidityDaysGT == 0 && len(c.CertSigAlgIn) == 0 && c.CertIsCA == nil &&
 		c.CertSelfSigned == nil && c.CertNotYetValid == nil && c.CertKeySizeLT == 0 &&
 		len(c.CertPubKeyAlgIn) == 0 && c.CertCAWithoutKeyCertSign == nil &&
-		c.CertKeyMismatch == nil &&
+		c.CertKeyMismatch == nil && c.CertChainBroken == nil &&
 		c.DuplicateLabel == nil && c.Orphan == nil &&
 		len(c.MechanismAnyOf) == 0 && len(c.MechanismMissing) == 0
 }